chore(dashboard): Abstract default pagination component
